Refactor, further abstract S3
