+++ server/commands/collectionExporter.ts
@@ -1,4 +1,5 @@
 import { Transaction } from "sequelize";
+import { v4 as uuidv4 } from "uuid";
 import {
   FileOperationFormat,
   FileOperationType,
@@ -6,7 +7,6 @@ import {
 } from "@shared/types";
 import { traceFunction } from "@server/logging/tracing";
 import { Collection, Event, Team, User, FileOperation } from "@server/models";
-import { getAWSKeyForFileOp } from "@server/storage/interfaces/s3";
 
 type Props = {
   collection?: Collection;
@@ -18,6 +18,11 @@ type Props = {
   transaction: Transaction;
 };
 
+function getKeyForFileOp(teamId: string, name: string) {
+  const bucket = "uploads";
+  return `${bucket}/${teamId}/${uuidv4()}/${name}-export.zip`;
+}
+
 async function collectionExporter({
   collection,
   team,
@@ -28,7 +33,7 @@ async function collectionExporter({
   transaction,
 }: Props) {
   const collectionId = collection?.id;
-  const key = getAWSKeyForFileOp(user.teamId, collection?.name || team.name);
+  const key = getKeyForFileOp(user.teamId, collection?.name || team.name);
   const fileOperation = await FileOperation.create(
     {
       type: FileOperationType.Export,
