+++ server/models/helpers/DocumentHelper.tsx
@@ -19,7 +19,7 @@ import { trace } from "@server/logging/tracing";
 import type Document from "@server/models/Document";
 import type Revision from "@server/models/Revision";
 import User from "@server/models/User";
-import { getSignedUrl } from "@server/storage/interfaces/s3";
+import FileStorage from "@server/storage/files";
 import diff from "@server/utils/diff";
 import parseAttachmentIds from "@server/utils/parseAttachmentIds";
 import Attachment from "../Attachment";
@@ -324,7 +324,10 @@ export default class DocumentHelper {
         });
 
         if (attachment) {
-          const signedUrl = await getSignedUrl(attachment.key, expiresIn);
+          const signedUrl = await FileStorage.getSignedUrl(
+            attachment.key,
+            expiresIn
+          );
           text = text.replace(
             new RegExp(escapeRegExp(attachment.redirectUrl), "g"),
             signedUrl
