+++ server/queues/tasks/UploadTeamAvatarTask.ts
@@ -1,6 +1,6 @@
 import { v4 as uuidv4 } from "uuid";
 import { Team } from "@server/models";
-import { uploadToS3FromUrl } from "@server/storage/interfaces/s3";
+import FileStorage from "@server/storage/files";
 import BaseTask, { TaskPriority } from "./BaseTask";
 
 type Props = {
@@ -20,7 +20,7 @@ export default class UploadTeamAvatarTask extends BaseTask<Props> {
       rejectOnEmpty: true,
     });
 
-    const avatarUrl = await uploadToS3FromUrl(
+    const avatarUrl = await FileStorage.uploadFileFromUrl(
       props.avatarUrl,
       `avatars/${team.id}/${uuidv4()}`,
       "public-read"
