+++ server/routes/api/attachments/attachments.ts
@@ -12,10 +12,7 @@ import { Attachment, Document, Event } from "@server/models";
 import AttachmentHelper from "@server/models/helpers/AttachmentHelper";
 import { authorize } from "@server/policies";
 import { presentAttachment } from "@server/presenters";
-import {
-  getPresignedPost,
-  publicS3Endpoint,
-} from "@server/storage/interfaces/s3";
+import FileStorage from "@server/storage/files";
 import { APIContext } from "@server/types";
 import { RateLimiterStrategy } from "@server/utils/RateLimiter";
 import { assertIn } from "@server/validation";
@@ -93,7 +90,7 @@ router.post(
       { transaction }
     );
 
-    const presignedPost = await getPresignedPost(
+    const presignedPost = await FileStorage.getPresignedPost(
       key,
       acl,
       maxUploadSize,
@@ -102,7 +99,7 @@ router.post(
 
     ctx.body = {
       data: {
-        uploadUrl: publicS3Endpoint(),
+        uploadUrl: FileStorage.getPublicEndpoint(),
         form: {
           "Cache-Control": "max-age=31557600",
           "Content-Type": contentType,
