+++ server/routes/api/fileOperations/fileOperations.test.ts
@@ -12,7 +12,7 @@ import { getTestServer } from "@server/test/support";
 
 const server = getTestServer();
 
-jest.mock("@server/storage/interfaces/s3");
+jest.mock("@server/storage/files");
 
 describe("#fileOperations.info", () => {
   it("should return fileOperation", async () => {
