+++ server/storage/files/__mocks__/index.ts
@@ -1,0 +1,9 @@
+export default {
+  uploadFile: jest.fn().mockReturnValue("/endpoint/key"),
+
+  getPublicEndpoint: jest.fn().mockReturnValue("http://mock"),
+
+  getSignedUrl: jest.fn().mockReturnValue("http://s3mock"),
+
+  getPresignedPost: jest.fn().mockReturnValue({}),
+};
