+++ server/storage/files/index.ts
@@ -1,0 +1,3 @@
+import S3Storage from "./S3Storage";
+
+export default new S3Storage();
